+++ lib/exfmt/ast/to_algebra.ex
@@ -111,7 +111,7 @@ defmodule Exfmt.Ast.ToAlgebra do
   def to_algebra({:%{}, _, contents}, ctx) do
     new_ctx = Context.push_stack(ctx, :map)
     body_doc = map_body_to_algebra(contents, new_ctx)
-    ["%{", nest(concat(break(""), body_doc), 2), break(""), "}"]
+    ["%{", nest(concat(break(""), body_doc), 2), break("", ","), "}"]
     |> concat()
     |> group()
   end